--- src/cai_lab_to_nwb/zaki_2024/__init__.py
@@ -1,2 +1,0 @@
-from .embargo_2024behaviorinterface import Embargo2024BehaviorInterface
-from .embargo_2024nwbconverter import Embargo2024NWBConverter
